Render volumes using extension settings (#508)
